--- src/main/java/webprogrammingTeam/matchingService/domain/Board/Dto/Request/BoardSaveRequest.java
@@ -1,7 +1,0 @@
-package webprogrammingTeam.matchingService.domain.Board.Dto.Request;
-
-import lombok.Builder;
-
-@Builder
-public record BoardSaveRequest(String title, String content){
-}
