--- src/main/java/webprogrammingTeam/matchingService/domain/Board/Dto/Response/BoardAllReadResponse.java
@@ -1,7 +1,0 @@
-package webprogrammingTeam.matchingService.domain.Board.Dto.Response;
-
-import lombok.Builder;
-
-@Builder
-public record BoardAllReadResponse(Long id, String title, String date) {
-}
