--- src/main/java/webprogrammingTeam/matchingService/domain/Board/Dto/Response/BoardIdReadResponse.java
@@ -1,9 +1,0 @@
-package webprogrammingTeam.matchingService.domain.Board.Dto.Response;
-
-import lombok.Builder;
-
-import java.util.List;
-
-@Builder
-public record BoardIdReadResponse(String tile, String date, String content, List<byte[]> imagesByte) {
-}
