--- src/main/java/webprogrammingTeam/matchingService/domain/Review/Dto/Request/ReviewSaveRequest.java
@@ -1,6 +1,0 @@
-package webprogrammingTeam.matchingService.domain.Review.Dto.Request;
-
-import lombok.Builder;
-@Builder
-public record ReviewSaveRequest(float rating, String content, String date){
-}
