+++ src/main/java/webprogrammingTeam/matchingService/domain/Review/dto/request/ReviewSaveRequest.java
@@ -1,0 +1,6 @@
+package webprogrammingTeam.matchingService.domain.Review.dto.request;
+
+import lombok.Builder;
+@Builder
+public record ReviewSaveRequest(float rating, String content, String date){
+}
