+++ src/main/java/webprogrammingTeam/matchingService/domain/Review/dto/response/ReviewAllReadResponse.java
@@ -1,0 +1,7 @@
+package webprogrammingTeam.matchingService.domain.Review.dto.response;
+
+import lombok.Builder;
+
+@Builder
+public record ReviewAllReadResponse(float rating, String content, String date) {
+}
