+++ src/main/java/webprogrammingTeam/matchingService/domain/Review/dto/response/ReviewIdReadResponse.java
@@ -1,0 +1,7 @@
+package webprogrammingTeam.matchingService.domain.Review.dto.response;
+
+import lombok.Builder;
+
+@Builder
+public record ReviewIdReadResponse(float rating, String content, String date){
+}
