+++ src/main/java/webprogrammingTeam/matchingService/domain/board/dto/request/BoardSaveRequest.java
@@ -1,0 +1,7 @@
+package webprogrammingTeam.matchingService.domain.board.dto.request;
+
+import lombok.Builder;
+
+@Builder
+public record BoardSaveRequest(String title, String content){
+}
