+++ src/main/java/webprogrammingTeam/matchingService/domain/board/dto/response/BoardAllReadResponse.java
@@ -1,0 +1,7 @@
+package webprogrammingTeam.matchingService.domain.board.dto.response;
+
+import lombok.Builder;
+
+@Builder
+public record BoardAllReadResponse(Long id, String title, String date) {
+}
