+++ src/main/java/webprogrammingTeam/matchingService/domain/board/dto/response/BoardIdReadResponse.java
@@ -1,0 +1,9 @@
+package webprogrammingTeam.matchingService.domain.board.dto.response;
+
+import lombok.Builder;
+
+import java.util.List;
+
+@Builder
+public record BoardIdReadResponse(String tile, String date, String content, List<byte[]> imagesByte) {
+}
