+++ src/main/java/webprogrammingTeam/matchingService/domain/board/repository/BoardRepository.java
@@ -1,0 +1,7 @@
+package webprogrammingTeam.matchingService.domain.board.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import webprogrammingTeam.matchingService.domain.board.entity.Board;
+
+public interface BoardRepository extends JpaRepository<Board, Long> {
+}
